+++ Makefile
@@ -1,6 +1,6 @@
 CC=gcc
 CFLAGS=-Wall -Werror -Wextra -pedantic -c -g
-OBJECTS = main.o alias.o argv.o env.o handlers.o input.o process.o strings.o
+OBJECTS = main.o alias.o argv.o env.o handlers.o input.o process.o strings.o shell.o
 all: hsh clean
 hsh: $(OBJECTS)
 	$(CC) $(OBJECTS) -o hsh
