+++ shell.c
@@ -1,0 +1,53 @@
+#include "shell.h"
+
+
+/**
+ * run_shell - entry point to shell
+ * Desc: run_shell function that executes hsh
+ * @env: double pointer to environment variables
+ * @errname: program name for error msg
+ * Return: returns a prompt to hsh shell v2
+ */
+int run_shell(char *errname, char **env)
+{
+	int exit_call = -1;
+	int proc_result = 0;
+
+	while (exit_call == -1)
+	{
+		cmdbuf_t *cmd = parse_command_buffer(proc_result, errname, env);
+		if (cmd->argv[0])
+		{
+			switch (check_handlers(cmd->argv[0]))
+			{
+			case 1:
+				exit_handler(cmd, &exit_call);
+				continue;
+			case 2:
+				change_dir(cmd);
+				continue;
+			case 3:
+				print_env(cmd, env);
+				continue;
+			}
+			cmd->argv[0] = parse_alias(cmd->argv[0]);
+		}
+		proc_result = forking(cmd);
+	}
+	return (proc_result);
+}
+
+cmdbuf_t *parse_command_buffer(int proc_result, char *errname, char **env)
+{
+    prompt();
+	input_t *input = get_input(proc_result);
+	cmdbuf_t *cmd = parse_input(input);
+
+	cmd->err_name = errname;
+	cmd->pre_alias = _strdup(cmd->argv[0]);
+	cmd->env = env;
+	free(input->buffer);
+	free(input);
+	parse_env(&cmd);
+	return (cmd);
+}
